+++ src/main/java/defencer/controller/entity/ProjectController.java
@@ -4,7 +4,9 @@ import com.jfoenix.controls.JFXButton;
 import com.jfoenix.controls.JFXComboBox;
 import com.jfoenix.controls.JFXDatePicker;
 import com.jfoenix.controls.JFXNodesList;
+import defencer.controller.AskFormController;
 import defencer.controller.CalendarController;
+import defencer.controller.MainActivityController;
 import defencer.controller.PremierLeagueController;
 import defencer.controller.update.UpdateProjectController;
 import defencer.data.ControllersDataFactory;
@@ -34,6 +36,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 
 import java.net.URL;
+import java.sql.SQLException;
 import java.time.LocalDate;
 import java.time.temporal.ChronoUnit;
 import java.util.List;
@@ -88,22 +91,28 @@ public class ProjectController implements Initializable {
     @FXML
     private JFXDatePicker dateFind;
 
+
     private ObservableList<Project> observableProjects = FXCollections.observableArrayList();
     private static final Long DEFAULT_PERIOD = 30L;
 
     @Override
     public void initialize(URL location, ResourceBundle resources) {
+
         insertProjectTable();
         loadProjects();
         comboProject.setItems(FXCollections
                 .observableArrayList(getProjectName()));
         comboProject.setValue("");
 
-        btnAddOneMore.setOnAction(this::newProject);
 
-        btnDelete.setOnAction(e -> deleteProject());
+        MainActivityController mainActivityController = (MainActivityController) ControllersDataFactory.getLink().get(MainActivityController.class, "class");
+        mainActivityController.showSmartToolbar();
 
-        btnUpdate.setOnMouseClicked(e -> loadProjects());
+        mainActivityController.getAddAction().setOnMouseClicked(this::newProject);
+        mainActivityController.getDeleteAction().setOnMouseClicked(e -> deleteProject());
+        mainActivityController.getUpdateAction().setOnMouseClicked(e -> loadProjects());
+        mainActivityController.getEditAction().setVisible(false);
+        mainActivityController.getPdfExportAction().setOnMouseClicked(e -> pdfReport());
 
         btnFind.setOnAction(e -> search());
 
@@ -113,7 +122,6 @@ public class ProjectController implements Initializable {
             }
         });
 
-        projectReport();
         projectConfigure();
     }
 
@@ -246,11 +254,10 @@ public class ProjectController implements Initializable {
 
     /**
      * Open new page to add one more project.
-     *
      * {@link SneakyThrows} here because i am totally sure that path to fxml is correct.
      */
     @SneakyThrows
-    private void newProject(ActionEvent event) {
+    private void newProject(MouseEvent event) {
 
         final FXMLLoader fxmlLoader = new FXMLLoader();
         fxmlLoader.setLocation(getClass().getResource("/entity/add/NewProject.fxml"));
@@ -373,9 +380,33 @@ public class ProjectController implements Initializable {
             NotificationUtil.warningAlert("Warning", "Select project first", NotificationUtil.SHORT);
             return;
         }
+
         try {
-            ServiceFactory.getProjectService().deleteEntity(project);
-            loadProjects();
+            final FXMLLoader fxmlLoader = new FXMLLoader();
+            fxmlLoader.setLocation(getClass().getResource("/askForm.fxml"));
+            Parent parent = fxmlLoader.load();
+
+            final Stage stage = new Stage();
+            Scene value = new Scene(parent);
+            value.getStylesheets().add("css/main.css");
+            stage.setScene(value);
+            stage.initModality(Modality.WINDOW_MODAL);
+            Window window = table.getScene().getWindow();
+            stage.initOwner(window);
+            stage.show();
+
+            stage.setOnHiding(event -> {
+                if ((boolean) ControllersDataFactory.getLink().get(AskFormController.class, "isDelete")) {
+
+                    try {
+                        ServiceFactory.getProjectService().deleteEntity(project);
+                    } catch (SQLException e) {
+                        e.printStackTrace();
+                    }
+                    loadProjects();
+                }
+            });
+
         } catch (Exception e) {
             NotificationUtil.errorAlert("Error", "Can't delete", NotificationUtil.SHORT);
         }
